edited model definitions to allow proper decimal precision for lat/lng

+++ models/food.js
@@ -26,11 +26,11 @@ module.exports = function (sequelize, DataTypes) {
             allowNull: false
         },
         latitude: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(17,14),
             allowNull: false
         },
         longitude: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(17,14),
             allowNull: false
         },
         familyFriendly: {
